requirements.txt
python-tlsh>=3.0.0
networkx>=2.5
tqdm>=4.60.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0

